backend | add flyway to the project

+++ src/main/java/com/elvira/customer/CustomerController.java
@@ -34,13 +34,13 @@ public class CustomerController {
 
     @DeleteMapping("/{customerId}")
     public void deleteCustomer (
-            @PathVariable("customerId") Integer customerId){
+            @PathVariable("customerId") Long customerId){
         customerService.deleteCustomerById(customerId);
     }
 
     @PutMapping("/{customerId}")
     public void updateCustomer(
-            @PathVariable("customerId") Integer customerId,
+            @PathVariable("customerId") Long customerId,
             @RequestBody CustomerUpdateRequest customerUpdateRequest
     ){
         customerService.updateCustomer(customerId, customerUpdateRequest);
