+++ src/main/java/com/elvira/customer/CustomerDao.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 
 public interface CustomerDao {
     List<Customer> selectAllCustomers();
-    Optional<Customer> selectCustomerById(Integer id);
+    Optional<Customer> selectCustomerById(Long id);
     void insertCustomer(Customer customer);
     boolean existsPersonWithEmail(String email);
-    boolean existsPersonWithId(Integer id);
-    void deleteCustomerById(Integer customerId);
+    boolean existsPersonWithId(Long id);
+    void deleteCustomerById(Long customerId);
     void updateCustomer(Customer update);
 }
