+++ src/main/java/com/elvira/customer/CustomerRepository.java
@@ -3,8 +3,8 @@ package com.elvira.customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository
-        extends JpaRepository<Customer, Integer> {
+        extends JpaRepository<Customer, Long> {
 
     boolean existsCustomerByEmail(String email);
-    boolean existsCustomerById(Integer id);
+    boolean existsCustomerById(Long id);
 }
