+++ src/main/java/com/elvira/customer/CustomerService.java
@@ -21,7 +21,7 @@ public class CustomerService {
         return customerDao.selectAllCustomers();
     }
 
-    public Customer getCustomer(Integer id) {
+    public Customer getCustomer(long id) {
         return customerDao.selectCustomerById(id)
                 .orElseThrow(() ->
                         new ResourceNotFoundException("customer with id [%s] not found".formatted(id))
@@ -42,14 +42,14 @@ public class CustomerService {
         );
     }
 
-    public void deleteCustomerById(Integer customerId) {
+    public void deleteCustomerById(Long customerId) {
         if (!customerDao.existsPersonWithId(customerId)) {
             throw new ResourceNotFoundException("customer with id [%s] does not exist".formatted(customerId));
         }
         customerDao.deleteCustomerById(customerId);
     }
 
-    public void updateCustomer(Integer customerId, CustomerUpdateRequest updateRequest) {
+    public void updateCustomer(Long customerId, CustomerUpdateRequest updateRequest) {
         Customer customer = getCustomer(customerId);
         boolean changes = false;
